// src/types/ChannelType.ts
export enum ChannelType {

  ECHO = '默语',
  SAY = '说话',
  YELL = '呼喊',
  SHOUT = '喊话',
  TELL = '悄悄话',
  REPLY = '回答',
  PARTY = '小队',
  ALLIANCE = '团队',
  FREECOMPANY = '部队',
  PVPTEAM = '战队',
  BEGINNER = '新人',
  EMOTE = '情感表情',

  LINKSHELL_1 = '通讯贝1',
  LINKSHELL_2 = '通讯贝2',
  LINKSHELL_3 = '通讯贝3',
  LINKSHELL_4 = '通讯贝4',
  LINKSHELL_5 = '通讯贝5',
  LINKSHELL_6 = '通讯贝6',
  LINKSHELL_7 = '通讯贝7',
  LINKSHELL_8 = '通讯贝8',

  CWLINKSHELL_1 = '跨服贝1',
  CWLINKSHELL_2 = '跨服贝2',
  CWLINKSHELL_3 = '跨服贝3',
  CWLINKSHELL_4 = '跨服贝4',
  CWLINKSHELL_5 = '跨服贝5',
  CWLINKSHELL_6 = '跨服贝6',
  CWLINKSHELL_7 = '跨服贝7',
  CWLINKSHELL_8 = '跨服贝8',

  ERROR = 'error',
  SYSTEM = 'system',

  SHELL_ERROR = 'shell_error',
  SHELL_SYSTEM = 'shell_system',

  HOLD = 'hold',

}